requirements.txt
certifi
chardet
idna
numpy
pdfminer.six
pycryptodome
requests
six
sortedcontainers
urllib3
python-dotenv
